serialized type generic
