+++ src/shapes/Object/types/index.ts
@@ -1,7 +1,7 @@
 import type { FabricObjectProps } from './FabricObjectProps';
 import type { TOptions } from '../../../typedefs';
 
-export type { SerializedObjectProps } from './SerializedObjectProps';
+export type { SerializedObjectProps } from './ObjectProps';
 export type { FabricObjectProps };
 
 export type TFabricObjectProps = TOptions<FabricObjectProps>;
